rentcrm/requirements.txt
dj-database-url==0.4.1
Django==1.10.2
python-dateutil==2.5.3
six==1.10.0
